requirements.txt
boto==2.49.0
certifi==2019.9.11
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
numpy==1.16.5
pandas==0.25.1
plotly==4.1.1
protobuf==3.6.0
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.3
pytz==2019.2
retrying==1.3.3
scikit-learn==0.21.2
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.15.5
wincertstore==0.2
